+++ lib/exception/sound_volume_view_exception.dart
@@ -1,0 +1,9 @@
+
+class SoundVolumeViewException implements Exception {
+  final String message;
+
+  SoundVolumeViewException(this.message);
+
+  @override
+  String toString() => 'SoundVolumeViewException: $message';
+}